\name{56773907-package}
\alias{56773907-package}
\alias{56773907}
\docType{package}
\title{\packageTitle{56773907}}
\description{\packageDescription{56773907}}
\details{
  The DESCRIPTION file: \packageDESCRIPTION{56773907}
  \packageIndices{56773907}

  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
  \packageAuthor{56773907}
  
  Maintainer: \packageMaintainer{56773907}
}
\references{
  This optional section can contain literature or other references for
  background information.
}
% Optionally other standard keywords, one per line,
% from the file KEYWORDS in the R documentation.
\keyword{package}
\seealso{
  Optional links to other man pages
}
\examples{
  ## Optional simple examples of the most important functions
  ## Use \dontrun{} around code to be shown but not executed
}
